SpeedReader/pyproject.toml
[tool.poetry]
name = "speedreader"
version = "0.1.0"
description = ""
authors = ["connorguy <connor_guy@icloud.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
spacy = "^3.3.0"
spacy-transformers = "1.1.6"
spacy-streamlit = "1.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
